feat: 4-tier copper currency + shop system wiring
Collapse the legacy split (Character.Gold/Silver/Platinum, PCData.BankGold)
into a single int64 copper field (Character.Coin, PCData.BankCoin). All
internal arithmetic now operates on copper; denomination split happens
only at I/O boundaries.

pkg/types/currency.go:
  - Constants: 10cp=1sp, 10sp=1gp, 10gp=1pp (so 1pp=1000cp).
  - CoinAmount + CoinFromCopper for denomination splits.
  - FormatCoin ("5p 3s 4c"), FormatCoinLong ("5 platinum, 3 silver, 4 copper").
  - ParseCoinDenom for parsing "5 gold", "100c", etc.

Object.Cost in area TOML is now interpreted as copper (no data edit — the
historical "gold" numbers are read as copper, restoring rough parity with
the original 100x-multiplied gold economy after the denomination collapse).

Save format: PlayerSave swaps {Gold, Silver, Platinum, BankGold} → {Coin,
BankCoin}. Loader test updated; old saves require migration (out of scope
here; flagged in .planning/STATE.md).

Bank commands (deposit/withdraw/balance) parse denominated input
("deposit 5 gold", "withdraw all", "deposit 100c"). Default denomination
for bare numbers is gold for back-compat ergonomics.

Shop system wired through server.LoadWorld: convert loader.ShopData →
shops.Shop, build ShopRegistry, install ShopHandler on the dispatcher so
buy/sell/list/value commands work. shops.GetSellPrice/GetBuyPrice apply
ProfitSell/ProfitBuy margins, haggle skill bonus, and a 1cp floor on
non-free items.

Death penalty (combat/damage.go) drops 10% of carried Coin (not bank);
corpse loot bundles all remaining coin into a single ItemTypeMoney with
denominated short_desc.

Sim test (pkg/loader/mobcoin_sim_test.go) reports per-level-bucket coin-
drop distribution (count, P10/median/P90, total world-pass injection) so
the economy baseline is locked before any rebalance.

rooms.toml (midgaard 3033): emergency newbie consumables added to the
general store so the basic cure potion + recall scroll are available
without trekking to the apothecary/magic shop.

+++ go/pkg/shops/shop.go
@@ -51,23 +51,23 @@ func (s *Shop) IsOpen(hour int) bool {
 	return hour >= s.OpenHour || hour < s.CloseHour
 }
 
-// GetBuyPrice calculates price shopkeeper pays to buy an item
-func (s *Shop) GetBuyPrice(obj *types.Object, ch *types.Character) int {
+// GetBuyPrice calculates price (in copper) shopkeeper pays for an item.
+func (s *Shop) GetBuyPrice(obj *types.Object, ch *types.Character) int64 {
 	if obj == nil {
 		return 0
 	}
 
-	// Base price
-	price := obj.Cost
+	// Base price (Object.Cost is now copper pieces)
+	price := int64(obj.Cost)
 
 	// Apply shop profit margin (lower = shop pays less)
-	price = price * s.ProfitBuy / 100
+	price = price * int64(s.ProfitBuy) / 100
 
 	// Apply haggle skill bonus
-	haggleBonus := getHaggleBonus(ch)
+	haggleBonus := int64(getHaggleBonus(ch))
 	price = price * (100 + haggleBonus) / 100
 
-	// Minimum price of 1
+	// Minimum price of 1cp on any non-free item
 	if price < 1 && obj.Cost > 0 {
 		price = 1
 	}
@@ -75,23 +75,23 @@ func (s *Shop) GetBuyPrice(obj *types.Object, ch *types.Character) int {
 	return price
 }
 
-// GetSellPrice calculates price customer pays to buy an item
-func (s *Shop) GetSellPrice(obj *types.Object, ch *types.Character) int {
+// GetSellPrice calculates price (in copper) the customer pays for an item.
+func (s *Shop) GetSellPrice(obj *types.Object, ch *types.Character) int64 {
 	if obj == nil {
 		return 0
 	}
 
-	// Base price
-	price := obj.Cost
+	// Base price (Object.Cost is now copper pieces)
+	price := int64(obj.Cost)
 
 	// Apply shop profit margin (higher = customer pays more)
-	price = price * s.ProfitSell / 100
+	price = price * int64(s.ProfitSell) / 100
 
 	// Apply haggle skill bonus (reduces price for good hagglers)
-	haggleBonus := getHaggleBonus(ch)
+	haggleBonus := int64(getHaggleBonus(ch))
 	price = price * (100 - haggleBonus/2) / 100
 
-	// Minimum price of 1
+	// Minimum price of 1cp on any non-free item
 	if price < 1 && obj.Cost > 0 {
 		price = 1
 	}
